+++ packages/fighting-design/_hooks/use-global/index.ts
@@ -20,7 +20,7 @@ export * from './interface.d'
 export const useGlobal = <T extends UseGlobalProp>(prop?: T): UseGlobalReturn => {
   /** 获取全局配置组件注入的依赖项 */
   const global: FightingGlobalProps | null = inject(FIGHTING_GLOBAL_PROPS_KEY, null)
-console.log(global)
+  console.log(global)
   /**
    * 获取组件的类型
    *